backend/requirements.txt
Django==5.2
djangorestframework==3.16.1
requests==2.31.0
django-cors-headers==4.3.1
django-environ==0.12.0
psycopg2-binary==2.9.11